requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyascii==0.3.3
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.3
coqpit-config==0.2.1
-e git+https://github.com/idiap/coqui-ai-TTS.git@9d03b314e35019b5ef97f3557d1ec929fe5042e4#egg=coqui_tts
coqui-tts-trainer==0.3.1
cycler==0.12.1
Cython==3.1.2
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
filelock==3.13.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.6.1
grpcio==1.74.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
huggingface-hub==0.34.3
idna==3.10
inflect==7.5.0
Jinja2==3.1.4
joblib==1.5.1
jsonlines==1.2.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.10.5
monotonic-alignment-search==0.2.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
networkx==3.3
num2words==0.5.14
numba==0.61.2
numpy==2.1.2
packaging==25.0
pillow==11.0.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pyparsing==3.2.3
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.54.1
typeguard==4.4.4
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
